refactor: URL handling and validation logic
url without protocol should be assumed valid for newsletter request
