+++ src/main/kotlin/fr/nicopico/n2rss/mail/newsletter/KotlinWeeklyNewsletterHandler.kt
@@ -23,7 +23,7 @@ import fr.nicopico.n2rss.mail.newsletter.jsoup.process
 import fr.nicopico.n2rss.models.Article
 import fr.nicopico.n2rss.models.Email
 import fr.nicopico.n2rss.models.Newsletter
-import fr.nicopico.n2rss.utils.toUrlOrNull
+import fr.nicopico.n2rss.utils.url.toUrlOrNull
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
 import org.jsoup.safety.Safelist
