+++ src/main/kotlin/fr/nicopico/n2rss/utils/url/UrlExt.kt
@@ -15,43 +15,13 @@
  * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-package fr.nicopico.n2rss.utils
+package fr.nicopico.n2rss.utils.url
 
-import jakarta.validation.Constraint
-import jakarta.validation.ConstraintValidator
-import jakarta.validation.ConstraintValidatorContext
-import jakarta.validation.Payload
-import kotlin.reflect.KClass
+import java.net.MalformedURLException
+import java.net.URL
 
-@MustBeDocumented
-@Constraint(validatedBy = [UrlValidator::class])
-@Target(
-    AnnotationTarget.FIELD,
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER,
-    AnnotationTarget.VALUE_PARAMETER,
-)
-@Retention(AnnotationRetention.RUNTIME)
-@Repeatable
-annotation class Url(
-    val message: String = "Invalid URL",
-    val groups: Array<KClass<*>> = [],
-    val payload: Array<KClass<out Payload>> = []
-)
-
-private val urlRegex = Regex(
-    "^((http|https):\\/\\/)?(www\\.)?([A-z]+)\\.([A-z]{2,})"
-)
-
-internal class UrlValidator : ConstraintValidator<Url, String> {
-    override fun isValid(
-        url: String?,
-        context: ConstraintValidatorContext?
-    ): Boolean {
-        if (url.isNullOrBlank()) {
-            return true
-        }
-        return urlRegex.matches(url)
-    }
+fun String.toUrlOrNull(): URL? = try {
+    URL(this)
+} catch (_: MalformedURLException) {
+    null
 }
